Direct console log to nothing during parsing
The parser outputs errors to the console during preprocessing. These errors are non essential and don't need to be there. As they are piped to the task output they needed to be silenced. 

Not the best way of doing it but i had no better solution at the moment. An issue #44 is open for help on it